progetto aggiornato e funzionante

+++ src/main/java/it/francesco/progetto/servicies/ProdottoInCarrelloService.java
@@ -3,6 +3,8 @@ package it.francesco.progetto.servicies;
 import it.francesco.progetto.entities.*;
 import it.francesco.progetto.repositories.*;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Isolation;
 import org.springframework.transaction.annotation.Transactional;
@@ -28,30 +30,35 @@ public class ProdottoInCarrelloService {
     @Autowired
     private AcquistoRepository acquistoRepository;
 
-
-    @Transactional(isolation = Isolation.REPEATABLE_READ)
-    public List<ProdottoInCarrello> getAllProdottiInCarrello() {
-        return prodottoInCarrelloRepository.findAll();
+    @Transactional(readOnly = true)
+    public ResponseEntity<List<ProdottoInCarrello>> getAllProdottiInCarrello() {
+        return new ResponseEntity<>(prodottoInCarrelloRepository.findAll(), HttpStatus.OK);
     }
 
-    @Transactional(isolation = Isolation.SERIALIZABLE)
-    public ProdottoInCarrello addProdottoInCarrello(ProdottoInCarrello p) {
+    @Transactional
+    public void addProdottoInCarrello(ProdottoInCarrello p) {
         Utente u = utenteRepository.findUtenteByUsername(p.getUtente().getUsername());
         Prodotto prodotto = prodottoRepository.findById(p.getProdotto().getId()).get();
         ProdottoInCarrello pIcPresente = prodottoInCarrelloRepository.findByProdottoAndUtente(prodotto, u);
-        ProdottoInCarrello pIc = new ProdottoInCarrello();
+        ProdottoInCarrello pIc = null;
 
         if(prodotto.getQta()>0)
             if (pIcPresente == null) {
+                pIc = new ProdottoInCarrello();
                 pIc.setProdotto(prodotto);
                 pIc.setQta(1);
                 pIc.setUtente(u);
                 pIc.setPrezzo(prodotto.getPrezzo());
                 prodottoInCarrelloRepository.save(pIc);
                 u.getProdottoInCarrello().add(pIc);
-                return pIc;
             } else {
                 if (pIcPresente.getQta() + 1 <= prodotto.getQta()) {
+
+                    pIcPresente.setQta(pIcPresente.getQta()+1);
+                    pIcPresente.setPrezzo(prodotto.getPrezzo()*pIcPresente.getQta());
+                    u.getProdottoInCarrello().add(pIcPresente);
+
+                    /*
                     pIc.setQta(pIcPresente.getQta() + 1);
                     pIc.setPrezzo(prodotto.getPrezzo() * pIc.getQta());
                     pIc.setUtente(u);
@@ -60,61 +67,58 @@ public class ProdottoInCarrelloService {
                     prodottoInCarrelloRepository.save(pIc);
                     u.getProdottoInCarrello().remove(pIcPresente);
                     u.getProdottoInCarrello().add(pIc);
-                    return pIc;
+                    */
                 }
-            return pIcPresente;
         }
-        return null;
     }
 
+    @Transactional
     public void deleteProdottoInCarrello(ProdottoInCarrello p) {
         prodottoInCarrelloRepository.delete(p);
     }
 
-    public ProdottoInCarrello findProdottoInCarrelloById(int id) {
-        return prodottoInCarrelloRepository.findById(id).get();
+    @Transactional(readOnly=true)
+    public ResponseEntity<ProdottoInCarrello> findProdottoInCarrelloById(int id) {
+        return new ResponseEntity<>(prodottoInCarrelloRepository.findById(id).get(), HttpStatus.OK);
     }
 
-    public List<ProdottoInCarrello> getProdottiInCarrelloByUsername(String username){
+    @Transactional(readOnly=true)
+    public ResponseEntity<List<ProdottoInCarrello>> getProdottiInCarrelloByUsername(String username){
         Utente u1 = utenteRepository.findUtenteByUsername(username);
-        System.out.println(u1.getProdottoInCarrello());
-        return u1.getProdottoInCarrello();
+        return new ResponseEntity<>(u1.getProdottoInCarrello(),HttpStatus.OK);
     }
 
-    @Transactional(isolation = Isolation.SERIALIZABLE)
-    public List<DettaglioOrdine> acquistaCarrello(String username) {
+    @Transactional
+    public void acquistaCarrello(String username) {
         Utente utente = utenteRepository.findUtenteByUsername(username);
         List<ProdottoInCarrello> list = utente.getProdottoInCarrello();
 
-        List<DettaglioOrdine> acquistabili = new LinkedList<>();
+        List<DettaglioOrdine> tmpDettaglioOrdine = new LinkedList<>();
         List<ProdottoInCarrello> acquistati = new LinkedList<>();
         Acquisto a = new Acquisto();
         a.setData(new Date());
 
-        System.out.println(list);
         for (ProdottoInCarrello p : list) {
             Prodotto pDB = prodottoRepository.findById(p.getProdotto().getId()).get();
-            pDB.setQta(pDB.getQta() - p.getQta());
+            if(pDB.getQta()>=p.getQta()){
+                pDB.setQta(pDB.getQta() - p.getQta());
 
-            DettaglioOrdine d = new DettaglioOrdine();
-            d.setAcquisto(a);
-            d.setQta(p.getQta());
-            d.setPrezzo(p.getPrezzo());
-            d.setProdotto(p.getProdotto());
+                DettaglioOrdine d = new DettaglioOrdine();
+                d.setAcquisto(a);
+                d.setQta(p.getQta());
+                d.setPrezzo(p.getPrezzo());
+                d.setProdotto(p.getProdotto());
+                tmpDettaglioOrdine.add(d);
+                acquistati.add(p);
 
-            acquistabili.add(d);
-            acquistati.add(p);
-
-            prodottoRepository.save(pDB);
-            dettaglioOrdineRepository.save(d);
-            prodottoInCarrelloRepository.delete(p);
+                dettaglioOrdineRepository.save(d);
+                prodottoInCarrelloRepository.delete(p);
+            }
         }
-
         list.removeAll(acquistati);
         a.setUtente(utente.getId());
-        a.setListaDettaglioOrdine(acquistabili);
+        a.setListaDettaglioOrdine(tmpDettaglioOrdine);
         utente.getLista().add(a);
         acquistoRepository.save(a);
-        return acquistabili;
     }
 }
